Renamed a few service modules - dropped _service from the names.
