+++ muckrake/tests/everything_runs_test.py
@@ -15,11 +15,11 @@
 from ducktape.tests.test import Test
 from ducktape.services.service_registry import ServiceRegistry
 
-from muckrake.services.zookeeper_service import ZookeeperService
+from muckrake.services.zookeeper import ZookeeperService
 from muckrake.services.kafka import KafkaService
 from muckrake.services.kafka_rest import KafkaRestService
 from muckrake.services.schema_registry import SchemaRegistryService
-from muckrake.services.register_schemas_service import RegisterSchemasService
+from muckrake.services.register_schemas import RegisterSchemasService
 
 
 class EverythingRunsTest(Test):
